requirements.txt
Django==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.12.1
psycopg2-binary==2.8.6
virtualenv==20.0.33